network: split out link_get_address_states()
No functional change, just refactoring and preparation for later
commits.
